[nova] auto-migrate from old versions

+++ openstack/nova/templates/bin/_db-migrate.tpl
@@ -3,27 +3,32 @@ set -e
 set -x
 
 {{- if hasPrefix "queens" .Values.imageVersion }}
-{{ if .Values.novaQueensUpgrade }}
-nova-manage cell_v2 map_cell0 --database_connection postgresql+psycopg2://{{ include "cell0_db_path" . }} || true
+NOVA_VERSION=$(nova-manage db version)
 
-virtualenv /tmp/nova-pike
-#curl -L https://github.com/openstack/nova/archive/16.1.6.tar.gz | tar -xzC /tmp/nova-pike
-git clone --depth 1 -b stable/pike https://github.com/openstack/nova.git /tmp/nova-pike/nova
-curl -L --output /tmp/nova-pike/upper-constraints.txt https://raw.githubusercontent.com/openstack/requirements/stable/pike/upper-constraints.txt
-source /tmp/nova-pike/bin/activate
-pip install -c /tmp/nova-pike/upper-constraints.txt -r /tmp/nova-pike/nova/requirements.txt /tmp/nova-pike/nova raven psycopg2
+if [ $NOVA_VERSION -eq '321' ]; then
 
-# Upgrade to pike
-nova-manage api_db sync --version 29 || true
-nova-manage db sync --version 362 || true
-echo 'Ignore errors due to online data migration'
-nova-manage db online_data_migrations
+    nova-manage cell_v2 map_cell0 --database_connection postgresql+psycopg2://{{ include "cell0_db_path" . }} || true
 
-# Create Cells if on pike, rerunning this doesn't break anything 
-nova-manage cell_v2 discover_hosts
-nova-manage cell_v2 simple_cell_setup
-deactivate
-rm -rf /tmp/nova-pike
+    virtualenv /tmp/nova-pike
+    #curl -L https://github.com/openstack/nova/archive/16.1.6.tar.gz | tar -xzC /tmp/nova-pike
+    git clone --depth 1 -b stable/pike https://github.com/openstack/nova.git /tmp/nova-pike/nova
+    curl -L --output /tmp/nova-pike/upper-constraints.txt https://raw.githubusercontent.com/openstack/requirements/stable/pike/upper-constraints.txt
+    source /tmp/nova-pike/bin/activate
+    pip install -c /tmp/nova-pike/upper-constraints.txt -r /tmp/nova-pike/nova/requirements.txt /tmp/nova-pike/nova raven psycopg2
+
+    # Upgrade to pike
+    nova-manage api_db sync --version 29 || true
+    nova-manage db sync --version 362 || true
+    echo 'Ignore errors due to online data migration'
+    nova-manage db online_data_migrations
+
+    # Create Cells if on pike, rerunning this doesn't break anything
+    nova-manage cell_v2 discover_hosts
+    nova-manage cell_v2 simple_cell_setup
+    deactivate
+    rm -rf /tmp/nova-pike
+
+fi
 {{ end }}
 {{- end }}
